+++ src/main/java/net/cupellation/api/CupellationEntrypoint.java
@@ -1,0 +1,6 @@
+package net.cupellation.api;
+
+public interface CupellationEntrypoint {
+
+    void registerCupellation();
+}