+++ src/main/java/net/cupellation/api/SmelterType.java
@@ -1,0 +1,6 @@
+package net.cupellation.api;
+
+import net.minecraft.block.Block;
+
+public record SmelterType(Block smelter, Block faucet, Block castingBasin, Block castingTable) {
+}