+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### Added:
 - smelter type max temperature
+- smelter block registration api
 ### Fixed:
 - 
 ### Changed: